Return lifecycle to resource and file element

+++ packages/ai-chat/src/browser/change-set-file-element.ts
@@ -14,14 +14,17 @@
 // SPDX-License-Identifier: EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0
 // *****************************************************************************
 
-import { MEMORY_TEXT, URI } from '@theia/core';
+import { DisposableCollection, Emitter, URI } from '@theia/core';
 import { inject, injectable, postConstruct } from '@theia/core/shared/inversify';
 import { ChangeSetElement, ChangeSetImpl } from '../common';
-import { createChangeSetFileUri } from './change-set-file-resource';
+import { ChangeSetFileResourceResolver, createChangeSetFileUri, UpdatableReferenceResource } from './change-set-file-resource';
 import { ChangeSetFileService } from './change-set-file-service';
+import { FileService } from '@theia/filesystem/lib/browser/file-service';
+import { ConfirmDialog } from '@theia/core/lib/browser';
 
 export const ChangeSetFileElementFactory = Symbol('ChangeSetFileElementFactory');
 export type ChangeSetFileElementFactory = (elementProps: ChangeSetElementArgs) => ChangeSetFileElement;
+type ChangeSetElementState = ChangeSetElement['state'];
 
 export const ChangeSetElementArgs = Symbol('ChangeSetElementArgs');
 export interface ChangeSetElementArgs extends Partial<ChangeSetElement> {
@@ -41,23 +44,74 @@ export interface ChangeSetElementArgs extends Partial<ChangeSetElement> {
 @injectable()
 export class ChangeSetFileElement implements ChangeSetElement {
 
+    static toReadOnlyUri(baseUri: URI, sessionId: string): URI {
+        return baseUri.withScheme('change-set-immutable').withAuthority(sessionId);
+    }
+
     @inject(ChangeSetElementArgs)
     protected readonly elementProps: ChangeSetElementArgs;
 
     @inject(ChangeSetFileService)
     protected readonly changeSetFileService: ChangeSetFileService;
 
-    protected _state: 'pending' | 'applied' | undefined;
+    @inject(FileService)
+    protected readonly fileService: FileService;
+
+    @inject(ChangeSetFileResourceResolver)
+    protected readonly resourceResolver: ChangeSetFileResourceResolver;
+
+    protected readonly toDispose = new DisposableCollection();
+    protected _state: ChangeSetElementState;
 
     protected originalContent: string | undefined;
 
+    protected readonly onDidChangeEmitter = new Emitter<void>();
+    readonly onDidChange = this.onDidChangeEmitter.event;
+    protected readOnlyResource: UpdatableReferenceResource;
+    protected changeResource: UpdatableReferenceResource;
+
     @postConstruct()
     init(): void {
+        this.getResources();
         this.obtainOriginalContent();
+        this.listenForOriginalFileChanges();
+        this.toDispose.push(this.onDidChangeEmitter);
     }
 
     protected async obtainOriginalContent(): Promise<void> {
         this.originalContent = await this.changeSetFileService.read(this.uri);
+        this.readOnlyResource.update({ contents: this.originalContent ?? '' });
+        if (this.state === 'applied') {
+
+        }
+    }
+
+    protected getResources(): void {
+        this.readOnlyResource = this.resourceResolver.tryGet(this.readOnlyUri) ?? this.resourceResolver.add(this.readOnlyUri, { autosaveable: false, readOnly: true });
+        let changed = this.resourceResolver.tryGet(this.changedUri);
+        if (changed) {
+            changed.update({ contents: this.targetState, onSave: content => this.writeChanges(content) });
+        } else {
+            changed = this.resourceResolver.add(this.changedUri, { contents: this.targetState, onSave: content => this.writeChanges(content), autosaveable: false });
+        }
+        this.changeResource = changed;
+        this.toDispose.pushAll([this.readOnlyResource, this.changeResource]);
+    }
+
+    protected listenForOriginalFileChanges(): void {
+        this.toDispose.push(this.fileService.onDidFilesChange(async event => {
+            if (!event.contains(this.uri)) { return; }
+            // If we are applied, the tricky thing becomes the question what to revert to; otherwise, what to apply.
+            const newContent = await this.changeSetFileService.read(this.uri).catch(() => '');
+            this.readOnlyResource.update({ contents: newContent });
+            if (newContent === this.originalContent) {
+                this.state = 'pending';
+            } else if (newContent === this.targetState) {
+                this.state = 'applied';
+            } else {
+                this.state = 'stale';
+            }
+        }));
     }
 
     get uri(): URI {
@@ -65,7 +119,7 @@ export class ChangeSetFileElement implements ChangeSetElement {
     }
 
     get readOnlyUri(): URI {
-        return this.elementProps.uri.withScheme(MEMORY_TEXT).withQuery(this.originalContent ?? '');
+        return ChangeSetFileElement.toReadOnlyUri(this.uri, this.elementProps.chatSessionId);
     }
 
     get changedUri(): URI {
@@ -84,13 +138,15 @@ export class ChangeSetFileElement implements ChangeSetElement {
         return this.changeSetFileService.getAdditionalInfo(this.uri);
     }
 
-    get state(): 'pending' | 'applied' | undefined {
+    get state(): ChangeSetElementState {
         return this._state ?? this.elementProps.state;
     }
 
-    protected set state(value: 'pending' | 'applied' | undefined) {
-        this._state = value;
-        this.elementProps.changeSet.notifyChange();
+    protected set state(value: ChangeSetElementState) {
+        if (this._state !== value) {
+            this._state = value;
+            this.onDidChangeEmitter.fire();
+        }
     }
 
     get type(): 'add' | 'modify' | 'delete' | undefined {
@@ -117,13 +173,16 @@ export class ChangeSetFileElement implements ChangeSetElement {
     }
 
     async apply(contents?: string): Promise<void> {
-        if (await this.changeSetFileService.trySave(this.changedUri)) { /** Continue */ } else if (this.type === 'delete') {
-            await this.changeSetFileService.delete(this.uri);
-            this.state = 'applied';
-        } else {
-            await this.writeChanges(contents);
+        if (!await this.confirm('Apply')) { return; }
+        if (!(await this.changeSetFileService.trySave(this.changedUri))) {
+            if (this.type === 'delete') {
+                await this.changeSetFileService.delete(this.uri);
+                this.state = 'applied';
+            } else {
+                await this.writeChanges(contents);
+            }
         }
-        this.changeSetFileService.closeDiffsFor(this.readOnlyUri);
+        this.changeSetFileService.closeDiff(this.readOnlyUri);
     }
 
     async writeChanges(contents?: string): Promise<void> {
@@ -132,6 +191,7 @@ export class ChangeSetFileElement implements ChangeSetElement {
     }
 
     async revert(): Promise<void> {
+        if (!await this.confirm('Revert')) { return; }
         this.state = 'pending';
         if (this.type === 'add') {
             await this.changeSetFileService.delete(this.uri);
@@ -140,7 +200,17 @@ export class ChangeSetFileElement implements ChangeSetElement {
         }
     }
 
+    async confirm(verb: string): Promise<boolean> {
+        if (this._state !== 'stale') { return true; }
+        await this.openChange();
+        const thing = await new ConfirmDialog({
+            title: `${verb} suggestion.`,
+            msg: `The file ${this.uri.path.toString()} has changed since this suggestion was created. Are you certain you wish to ${verb.toLowerCase()} the change?`
+        }).open(true);
+        return !!thing;
+    }
+
     dispose(): void {
-        this.changeSetFileService.closeDiffsFor(this.readOnlyUri);
+        this.toDispose.dispose();
     }
 }
